requirements.txt
anyio==4.2.0
asgiref==3.6.0
azure-core==1.29.6
azure-storage-blob==12.19.0
bleach==6.0.0
boto3==1.26.90
botocore==1.29.90
cachetools==5.3.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
cloudinary==1.39.0
cryptography==41.0.7
Django==4.1.7
django-ckeditor==6.7.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-js-asset==2.2.0
django-storages==1.13.2
django-summernote==0.8.20.0
django-tinymce==3.6.0
djangorestframework==3.14.0
dnspython==2.4.2
exceptiongroup==1.2.0
google-api-core==2.11.0
google-auth==2.17.3
google-cloud-core==2.3.2
google-cloud-secret-manager==2.17.0
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.13.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==20.1.0
idna==3.4
isodate==0.6.1
jmespath==1.0.1
mssql-django==1.3
packaging==23.2
Pillow==9.4.0
proto-plus==1.23.0
protobuf==4.22.3
psycopg2==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pymongo==4.6.1
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.2.4
stripe==7.0.0
typing_extensions==4.8.0
tzdata==2022.7
urllib3==1.26.15
webencodings==0.5.1
whitenoise==6.4.0
